design reorg
