+++ exps/quora-eda/elmo_classification_GLUE.py
@@ -1,28 +1,4 @@
-#########################################################
-################### GLOBALS & IMPORTS ###################
-#########################################################
-import time
-day = time.strftime(f'%Y-%m-%d')
 
-NUM_EXAMPLES = 20000 #20000 #100000  404290   #EXP CHANGE 1
-EVAL_METHOD = 'GLUE' #base #GLUE
-METHOD = f'SIZE_{NUM_EXAMPLES//1000}k_{EVAL_METHOD}' #100k #20k  #full  #50-50  #no-repeat  #filter
-TITLE = f'quora-{METHOD}_{day}'
-
-MODEL_NAME = "ELMo-allennlp"
-MAX_LENGTH = 40 #elmo_change # *2 FOR BERT multiply by 2 because now passing two paraphrases together
-
-TRAIN_SIZE = 0.8
-TEST_SIZE = 0.2
-RNDM_SEED = 42
-STRAT = False #EXP CHANGE 2
-CHECK_DUP = False #EXP CHANGE 3
-
-LR = 1e-4
-NUM_EPOCHS = 6 # elmo_change
-BATCH_SIZE = 512
-DROP_LAST = False
-PUNC = '!"#$%&()*+,-./:;<=>?@[\\]^_`{|}~—“”'
 
 
 # WandB parameters
@@ -121,445 +97,5 @@ warnings.filterwarnings("ignore")
 ################# DATASET CLASS W/ ELMO #################
 #########################################################
 
-class QuoraDatasetElmo(Dataset):
-    def __init__(self, para_dataset: str = 'quora', model_name: str = 'ELMo-allennlp', 
-                 num_examples: int = 20000, max_length: int = 40, 
-                 train_size: float = 0.8, test_size: float = 0.2,seed: int = None):
-        
-        self.para_dataset = para_dataset
-        self.model_name = model_name
-        self.num_examples = num_examples
-        self.max_length = max_length
-        self.seed = seed # set to None for random
 
-        self.sentence_id_pair_list = [] # index of this list corresponds to index of tokenized_sentences
-        # elmo_change
-        # self.tokenized_sentences = [] # lists of [CLS] <s1> [SEP] <s2> [SEP]; where <s1> and <s2> are tokenized sentences
-        self.s1_tokenized = [] # lists of sentences of token_ids from s1
-        self.s2_tokenized  = [] # lists of sentences of token_ids from s2
-        self.labels = [] # list of labels as ints: 0,1
 
-        if CHECK_DUP:
-            self.sent_id_list = []
-            self.dup = 0
-
-        #elmo_change
-        # https://github.com/luismsgomes/mosestokenizer
-        self.tokenizer = MosesTokenizer('en') 
-        
-
-
-        self.bad_words = ["-LSB-", "\\", "``", "-LRB-", "????", "n/a", "'"] #, "//" #js add if you want to filter out some URLs
-
-        # load quora, mrpc, etc.
-        print(f'Loading {para_dataset} dataset...\n')
-        self._load_dataset()
-        self._train_test_split(train_size, test_size, STRAT)
-        self.split('train')
-
-    def split(self, train_or_test: str = 'train'):
-        self._split = train_or_test
-        print(f'{self.para_dataset} split set to {self._split}')
-
-    def __len__(self) -> int:
-        if self._split == 'train':
-            return len(self.train_labels) 
-
-        if self._split == 'test':
-            return len(self.test_labels)
-        # return len(self.labels) #changed from ParaDataset
-
-    # based on Shi's gen_batch
-    def __getitem__(self, idx: int) -> Tuple[Tensor, Tensor, Tensor]:
-        if self._split == 'train':
-            # self.train_sents has shape B x 2 x self.max_length x word_length==50
-            return self.train_sents[idx,:,:,:], torch.tensor(self.train_labels[idx])
-
-            # return torch.tensor(self.train_sents[idx].squeeze()), torch.tensor(self.train_labels[idx]) 
-            # #elmo_change ^^ had to squeeze() to pass shape [T,50] instead of [1,T,50] for batching
-
-        if self._split == 'test':
-            # self.test_sents has shape B x 2 x self.max_length x word_length==50
-
-            return self.test_sents[idx,:,:,:], torch.tensor(self.test_labels[idx]) 
-
-            # return torch.tensor(self.test_sents[idx].squeeze()), torch.tensor(self.test_labels[idx]) 
-            # #elmo_change ^^ had to squeeze() to pass shape [T,50] instead of [1,T,50] for batching
-
-        # return torch.tensor(self.tokenized_sentences[idx].ids), torch.tensor(self.labels[idx])
-    
-    # TODO add MRPC, PAN...
-    def _load_dataset(self):
-        """
-        Loads relevant paraphrase dataset based on argument passed to constructor and creates attributes for model training.
-        """
-
-        # load quora from HuggingFace
-        if self.para_dataset == 'quora':
-            quora = datasets.load_dataset(self.para_dataset)
-            
-            quora_shuffled = quora.shuffle(seed=self.seed) #js holdover from ParaDataset; ensures if you're using less than total dataset, you are at least shuffling it first.
-            count = 0
-            if STRAT:
-                pos_label_count = 0
-                neg_label_count = 0
-                print('STRAT==True')
-
-            bad_count = 0              #js added for testing
-            sentence_id_pair_list = [] #js added for testing
-            label_list = []            #js added for testing
-
-            for pair in quora_shuffled['train']:
-                # count += 1 #NOTE:  incrementing here means you don't actually get 20k examples. It's what Shi did.
-                if count >= self.num_examples:
-                    break
-
-                label = pair['is_duplicate']
-                s1_id = pair['questions']['id'][0]
-                s2_id = pair['questions']['id'][1]
-                s1 = pair['questions']['text'][0]
-                s2 = pair['questions']['text'][1]
-                
-                exist_bad_word = False
-                for i in self.bad_words:
-                    if (i in s1 or i in s2):
-                        exist_bad_word = True
-
-                if exist_bad_word: 
-                    bad_count += 1  #js added for testing
-                    continue # skip pair if it contains element from self.bad_words
-
-                #js experiment 50-50 split
-                if STRAT:
-                    if label:
-                        pos_label_count +=1
-                        if pos_label_count > self.num_examples//2:
-                            continue #if half of examples are positive, go to top of loop to find more negative
-                    else:
-                        neg_label_count +=1
-                        if neg_label_count > self.num_examples//2:
-                            continue #if half of examples of negative, go to top of loop to find more positive
-
-                #js experiment no duplicate questions
-                
-                if CHECK_DUP:
-                    if s1_id in self.sent_id_list:
-                        self.dup += 1
-                        continue #if either sentence is in the set, go to top of loop
-                    elif s2_id in self.sent_id_list:
-                        self.dup += 1
-                        continue
-                    else:  #otherwise add to sent_id_list
-                        self.sent_id_list.append(s1_id) 
-                        self.sent_id_list.append(s2_id)
-
-                #### CODE FOR TESTING HOW MANY EXAMPLES ARE POSITIVE AND NEGATIVE ####
-                self.sentence_id_pair_list.append((s1_id,s2_id)) #js added
-                label_list.append(label)                    #js added
-                #### CODE FOR TESTING HOW MANY EXAMPLES ARE POSITIVE AND NEGATIVE ####
-
-                #elmo_change thru `count += 1` 
-                # tokenize the sentences and turn into tensor of ELMo character_ids
-                # concat_sents = ' '.join([s1,s2])
-                # print(concat_sents)
-                # interim_tokenized = batch_to_ids([self.tokenizer.tokenize(concat_sents, escape=False)]) # shape: (1,seq length, 50)
-                s1_interim_tokenized = batch_to_ids([self.tokenizer(s1)]) # shape: (1,seq length, 50)
-                s2_interim_tokenized = batch_to_ids([self.tokenizer(s2)]) # shape: (1,seq length, 50) 
-
-                # add padding to self.max_length && truncate if number of tokens > self.max_length
-                # # full_tokenized = torch.zeros(1, self.max_length, 50, dtype=torch.int64) # shape: (one example, self.max_length, 50 for ELMo character encoding)
-                s1_full_tokenized = torch.zeros(1, self.max_length, 50, dtype=torch.int64)
-                s2_full_tokenized = torch.zeros(1, self.max_length, 50, dtype=torch.int64)
-
-                # Truncate to self.max_length
-                # # full_tokenized[:,:interim_tokenized.shape[1],:] = torch.clone(interim_tokenized[:,self.max_length,:]) #previously a BUG for 2x WandB ELMo runs; THERE IS NO `:` before self.max_length...so it just copied the self.max_length cell across the whole tensor
-                # s1_full_tokenized[:,:s1_interim_tokenized.shape[1],:] = torch.clone(s1_interim_tokenized[:,:self.max_length,:])
-                # s2_full_tokenized[:,:s2_interim_tokenized.shape[1],:] = torch.clone(s2_interim_tokenized[:,:self.max_length,:])
-                s1_full_tokenized[:,:s1_interim_tokenized.shape[1],:] = torch.clone(s1_interim_tokenized[:,:self.max_length,:])
-                s2_full_tokenized[:,:s2_interim_tokenized.shape[1],:] = torch.clone(s2_interim_tokenized[:,:self.max_length,:])               
-                
-                # Ensure values and shapes match
-                assert torch.all(torch.eq(s1_full_tokenized[:,:s1_interim_tokenized.shape[1],:], s1_interim_tokenized[:,:self.max_length,:]))
-                assert torch.all(torch.eq(s2_full_tokenized[:,:s2_interim_tokenized.shape[1],:], s2_interim_tokenized[:,:self.max_length,:]))
-                assert s1_full_tokenized.shape == (1,self.max_length,50)
-                assert s2_full_tokenized.shape == (1,self.max_length,50)
-
-                # self.tokenized_sentences.append(full_tokenized)
-                self.s1_tokenized.append(s1_full_tokenized)
-                self.s2_tokenized.append(s2_full_tokenized)      
-
-
-                count += 1  # NOTE: count should be incremented here to get 20k examples
-
-
-            #### PRINT STATEMENTS FOR TESTING ####
-            #print(f'Total time: {(time.time() - t0):.2f} seconds')
-            print('\n#### PRINT STATEMENTS FOR TESTING ####')
-            print(f'Total pairs of questions: {len(self.sentence_id_pair_list)}')
-            print(f'`count` is {count}, should be {self.num_examples}')
-            print(f'`bad_count` is {bad_count}')
-            print(f'Number of paraphrases = {sum(label_list)}, {sum(label_list)/len(self.sentence_id_pair_list)*100:.2f}% of total')
-            print(f'Number of NON-paraphrases = {len(label_list) - sum(label_list)}, {(len(label_list) - sum(label_list))/len(self.sentence_id_pair_list)*100:.2f}% of total')
-            print('#### PRINT STATEMENTS FOR TESTING ####\n')
-            #### PRINT STATEMENTS FOR TESTING ####
-
-            # map boolean list of labels to list of 0,1 ints
-            self.labels = list(map(int, label_list))
-            if STRAT:
-              print(f'pos_label_count = {pos_label_count}')
-              print(f'neg_label_count = {neg_label_count}')
-
-
-    def _train_test_split(self, train_size: float =0.8, test_size: float = 0.2, strat=False):
-        """
-        If called, creates train/test splits of sents/labels.
-        """
-        # q = torch.stack((torch.cat(s1_tokenized),torch.cat(s2_tokenized)),dim=1)
-        # SHAPE: num_examples x (s1,s2) x max_length x char length (50)
-        # SHAPE: num_examples x 2       x 40         x 50
-        self.tokenized_sentences = torch.stack((torch.cat(self.s1_tokenized), 
-                                                torch.cat(self.s2_tokenized)), 
-                                                dim=1)
-        try:
-            assert self.tokenized_sentences.shape == (self.num_examples,2,40,50)
-        except AssertionError:
-            print(f'self.tokenized_sentences.shape: {self.tokenized_sentences.shape}\nintended shape: {(self.num_examples,2,40,50)}')
-
-
-        if strat:
-            self.train_sents, self.test_sents, self.train_labels, self.test_labels = train_test_split(
-                self.tokenized_sentences, self.labels, train_size=train_size, test_size=test_size, 
-                random_state=self.seed, shuffle = True, stratify = self.labels)
-
-        else: 
-            self.train_sents, self.test_sents, self.train_labels, self.test_labels = train_test_split(
-                self.tokenized_sentences, self.labels, train_size=train_size, test_size=test_size, 
-                random_state=self.seed, shuffle = True, stratify = None
-            )
-        print(f'Training Size: {len(self.train_sents)}')
-        print(f'Testing Size: {len(self.test_sents)}')
-
-
-
-#########################################################
-################## DATASET & DATALOADER #################
-#########################################################
-
-quora = QuoraDatasetElmo(para_dataset = 'quora', model_name = MODEL_NAME, 
-                     num_examples = NUM_EXAMPLES, max_length = MAX_LENGTH,
-                     train_size = TRAIN_SIZE, test_size = TEST_SIZE, seed = RNDM_SEED)
-
-# js NOTE: HAD TO COPY QUORA BECAUSE CHANGING THE SPLIT CHANGED THE DATALOADERS.
-quora.split('train')
-train_quora = copy.copy(quora)
-train_dataloader = DataLoader(train_quora, batch_size=BATCH_SIZE, shuffle=True, drop_last=DROP_LAST, pin_memory=True)
-quora.split('test')
-test_quora = copy.copy(quora)
-test_dataloader = DataLoader(test_quora, batch_size=BATCH_SIZE, shuffle=True, drop_last=DROP_LAST, pin_memory=True)
-
-print('\n***CHECK DATASET LENGTHS:***')
-print(f'len(train_dataloader.dataset) = {len(train_dataloader.dataset)}')
-print(f'len(test_dataloader.dataset) = {len(test_dataloader.dataset)}')
-
-print('\n***CHECK self.tokenized_sentences.shape: ')
-print(f'quora.tokenized_sentences.shape: {quora.tokenized_sentences.shape}\nintended shape: {(quora.num_examples,2,40,50)}')
-#########################################################
-################## ElmoClassifier Class #################
-#########################################################
-
-
-class ElmoClassifier(torch.nn.Module):
-
-    def __init__(self, options_file: str, weight_file: str, num_output_representations: int, requires_grad: bool=True, dropout: float =0):
-        super().__init__()
-        self.elmo = Elmo(options_file = options_file, weight_file = weight_file, 
-                         num_output_representations = num_output_representations, requires_grad=requires_grad, dropout=dropout)
-        self.elmo_hidden_size = self.elmo.get_output_dim()
-        self.linear1 = torch.nn.Linear(in_features=self.elmo_hidden_size*3 + 1, out_features=512) #TODO change out_features to 50 for simplicity?
-        self.relu = torch.nn.ReLU()
-        self.linear2 = torch.nn.Linear(in_features=512, out_features=1)
-
-    def forward(self, sents: torch.Tensor) -> torch.Tensor:
-            # sents has shape B x 2 x self.max_length x word_length==50
-            #    where dim=1 is (sent1,sent2) which is why we index as follows
-        # try:
-        #     assert sents.shape == (BATCH_SIZE,2,40,50)
-        # except AssertionError:
-        #     print(f'sents.shape: {sents.shape}\nintended shape: {(BATCH_SIZE,2,40,50)}')
-        B = sents.shape[0] # batch_size because drop_last =False
-        assert sents.shape == (B,2,40,50)
-        sent_1 = sents[:,0,:,:] # shape: B x 1 x MAX_LENGTH==40 x 50
-        sent_2 = sents[:,1,:,:] # shape: B x 1 x MAX_LENGTH==40 x 50
-        
-        elmo_sent_1 = self.elmo(sent_1)  # shape: B x MAX_LENGTH==40 x 1024
-        elmo_sent_2 = self.elmo(sent_2)
-        average_sent_1 = elmo_sent_1['elmo_representations'][0].mean(dim=1) #B x 1024
-        average_sent_2 = elmo_sent_2['elmo_representations'][0].mean(dim=1) #B x 1024
-
-        u,v = average_sent_1, average_sent_2  
-        assert u.shape == (B,1024)
-        assert v.shape == (B,1024)
-
-        # FROM GLUE baseline: https://github.com/nyu-mll/GLUE-baselines/blob/b1c82396d960fd9725517089822d15e31b9882f5/src/models.py#L330
-        # return torch.cat([s1_enc, s2_enc, torch.abs(s1_enc - s2_enc), s1_enc * s2_enc], 1)
-        cat_vect = torch.cat((u, v, torch.norm(u-v, p=1, dim=1, keepdim=True), u*v), dim=1)  #NOTE: 2022-01-26; changed to p=1 after looking at GLUE baseline code
-        assert cat_vect.shape == (u.shape[0],u.shape[1]*3+1) # shape [batch, 1024*3+1]
-        x = self.linear1(cat_vect)
-        x = self.relu(x)
-        logits = self.linear2(x)
-        assert logits.shape == (B,1)
-        return logits
-    
-    #TODO:
-    # 1. add a cosine similarity comparison between two sentences
-
-
-    # https://github.com/allenai/allennlp/blob/main/allennlp/modules/elmo.py
-
-
-#########################################################
-###################### TRAINING LOOP ####################
-#########################################################
-
-
-# WandB init and config (based on argument dictionaries in imports/globals cell)
-wandb.init(**wandb_init_dict)
-wandb.config = wandb_config_dict
-
-# def compute_n_correct(preds, targets):
-#     return torch.sum((torch.sigmoid(preds.squeeze()).round() == targets)).cpu().item()
-
-# train on gpu if availble, set `device` as global variable
-device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-
-torch.manual_seed(RNDM_SEED)
-
-
-model = ElmoClassifier(options_file = OPTIONS_FILE,
-                       weight_file = WEIGHT_FILE, 
-                       num_output_representations = 1, 
-                       requires_grad=True, 
-                       dropout=0)
-
-model.to(device)
-
-optimizer = optim.Adam(model.parameters(), lr=LR)
-loss_fn = BCEWithLogitsLoss()
-
-train_losses = []
-test_losses = []
-train_accs = []
-test_accs = []
-train_f1 = []
-test_f1 = []
-train_precision = []
-test_precision = []
-train_recall = []
-test_recall = []
-
-
-t0 = time.time()
-for epoch in range(NUM_EPOCHS):
-    running_loss = 0
-    n_correct = 0
-    preds_accum = []
-    targets_accum = []
-
-    for batch, targets in tqdm(train_dataloader, leave=False):
-    # for batch, targets in train_dl:
-        targets_accum.append(targets)
-        batch, targets = batch.to(device), targets.to(device)
-        preds = model(batch)
-
-        loss = loss_fn(preds.squeeze(), targets.float())
-
-        preds_sigmoid = torch.sigmoid(preds.squeeze())
-        preds_accum.append(preds_sigmoid.cpu().detach().numpy().round()) 
-        
-        loss.backward()
-        optimizer.step()
-        optimizer.zero_grad()
-
-        running_loss += loss.cpu().item()
-
-
-    # https://discuss.pytorch.org/t/calculating-f1-score-over-batched-data/83348/2    
-    preds_accum = np.concatenate(preds_accum)
-    targets_accum = np.concatenate(targets_accum)
-    epoch_f1 = f1_score(targets_accum, preds_accum, average='binary')
-    epoch_precision = precision_score(targets_accum, preds_accum, average='binary')
-    epoch_recall = recall_score(targets_accum, preds_accum, average='binary')
-    epoch_accuracy = accuracy_score(targets_accum, preds_accum)
-    
-
-    train_losses.append(running_loss / len(train_dataloader.dataset))
-    train_accs.append(epoch_accuracy)
-    train_f1.append(epoch_f1)
-    train_precision.append(epoch_precision)
-    train_recall.append(epoch_recall)
-
-    t1 = time.time()
-    print("="*20)
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Train Loss: {train_losses[-1]:>0.5f}")
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Train Accuracy: {train_accs[-1]*100:>0.2f}%" )
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Train F1: {train_f1[-1]*100:>0.2f}" )
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Train Prec: {train_precision[-1]:>0.2f}" )
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Train Recall: {train_recall[-1]:>0.2f}" )
-
-
-    running_loss = 0
-    n_correct = 0
-    preds_accum = []
-    targets_accum = []
-
-    with torch.no_grad():
-        for batch, targets in tqdm(test_dataloader, leave=False):
-        # for batch, targets in test_dl:
-            targets_accum.append(targets)
-            batch, targets = batch.to(device), targets.to(device)
-            preds = model(batch)
-            loss = loss_fn(preds.squeeze(), targets.float())
-            preds_sigmoid = torch.sigmoid(preds.squeeze())
-            preds_accum.append(preds_sigmoid.cpu().round())
-
-            running_loss += loss.cpu().item()
-
-
-    preds_accum = np.concatenate(preds_accum)
-    targets_accum = np.concatenate(targets_accum)
-    epoch_f1 = f1_score(targets_accum, preds_accum, average='binary')
-    epoch_precision = precision_score(targets_accum, preds_accum, average='binary')
-    epoch_recall = recall_score(targets_accum, preds_accum, average='binary')
-    epoch_accuracy = accuracy_score(targets_accum, preds_accum)
-    
-    test_losses.append(running_loss / len(test_dataloader.dataset))
-    test_accs.append(epoch_accuracy)
-    test_f1.append(epoch_f1)
-    test_precision.append(epoch_precision)
-    test_recall.append(epoch_recall)
-
-    t1 = time.time()
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Test Loss: {test_losses[-1]:>0.5f}")
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Test Accuracy: {test_accs[-1]*100:>0.2f}%" )
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Test F1: {test_f1[-1]*100:>0.2f}" )
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Test Prec: {test_precision[-1]:>0.2f}" )
-    print(f"Epoch {epoch+1}/{NUM_EPOCHS} Test Recall: {test_recall[-1]:>0.2f}" )
-    
-    print(f"\nEpoch {epoch+1}/{NUM_EPOCHS} Total EPOCH Time: {(t1-t0)/60:>0.2f} min")
-    
-
-    # WandB log loss, accuracy, and F1 at end of each epoch
-    wandb.log({"Train/Loss": train_losses[-1],
-               "Train/Acc": train_accs[-1],
-               "Train/F1": train_f1[-1],
-               "Train/Recall": train_recall[-1],
-               "Train/Precision": train_precision[-1],
-               "Test/Loss": test_losses[-1],
-               "Test/Acc": test_accs[-1],
-               "Test/F1": test_f1[-1],
-               "Test/Recall": test_recall[-1],
-               "Test/Precision": test_precision[-1],
-               "Epoch Time": (t1-t0)/60})
-
-    t0 = time.time()
